wrap ifstream with custom line reader
Wrap a basic input file stream with a line reader class. The class
should eventually handle all the edge cases that occur while reading a
file line by line, and only return a full line regardless of the length
of the line. Possibly checking the bits of the file stream and appending
to the existing string each time.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>